// generated/src/app/api-models/hold-reason.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class HoldReason {
  holdReasonSequence: number;
  seqClaimId: number;
  lineNumber: number;
  subLineCode: string;
  history: string;
  reasonCode: string;
  sourceId: string;
  insertUser: string;
  insertDatetime: Date;
  insertProcess: string;
  updateUser: string;
  updateDatetime: Date;
  updateProcess: string;
  securityCode: string;
}